New camera setup
Tiago pcl tutorial
-Changed corner names according to new camera reference orientation
-Publish corner markers into a unique topic of type MarkerArray (/segment_table/corners)
-Commented (to delete) the part where computes the grasp point and angle (now done in iri_demos_kinova)

Iri_demos_kinova
-Subscribe to the corners topic (corners_callback)-> this function now does what before was done in tiago_pcl_tutorial segmCloudCallback:
	-Transforms corner point wrt base link
	-Computes edges sizes
	-Computes the grasp point and edge angle with corners wrt base link
-Changed garment_angle_callback function to an internal one (not a callback, since now the angle is computed in iri_demos_kinova)
-SM: Centred the ROTATE_POST_GRASP pose to the new camera's position

TODO:
-iri_demos_kinova: Get name of corners based on base_link (now obtained correctly based in camera ref)
-tiago_pcl_tutorial:
	-Separate point cloud in Pick and Place zones
	-Make table segmentation threshold configurable
	-Obtain corners robustly (filter noise point from nontable pointcloud)

Camera setup info: The camera has been rotated in its x axis, so the previous nomenclature of the corners does not correspond to the base_link reference (e.g. pt_up_left is now pt_down_right). In consequence, the grasp point that its selecting its on the farthest longest edge.
After changing the corners name so it makes sense with the base_link, the grasp point is an incorrect one (due to the + and - of axis y used to compute it).
Instead of recalculating it wrt the camera reference, I will move all the grasp point computation to the demo node, so it is always wrt the robot reference and does not depend on the camera's orientation.
Therefore, the tiago_pcl_tutorial will be in charge of the first processing of the pointcloud (filtering, segmentation, corner detection, min point to know pile height) but not of the specific computations related to the demo (edge and grasp point selection).