Deformation through depth OpenCV
Segmented the depth image using the image_difference package.
-It uses the color image (obtained from ros topic) to segment the garment (did not manage to segment directly from the depth image).
-The computed mask is applied to the depth image (obtained from ros topic) to segment the garment.

-/deformation folder includes first python and cpp tests. Not functional

Goal: Get the depth mean of squares of the folded cloth and compare it to the initial depth. This information will be used to determine if the cloth is wrinkled/bent or not
Steps: Use OpenCV and python to speed up the implementation to test if it works for our goal

TODO: Divide garment in sections. Interpret depth values. Calculate mean